// src/generated/api/reporting/reporting.schemas.ts
export type ReportGranularity = (typeof ReportGranularity)[keyof typeof ReportGranularity];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ReportGranularity = {
  DAY: 'DAY',
  MONTH: 'MONTH',
  QUARTER: 'QUARTER',
  YEAR: 'YEAR',
} as const;